Remove entire no italics file and just added one test case to the original test file for ChangelogParser
